% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-characters.R
\docType{data}
\name{characters}
\alias{characters}
\title{A morphological phylogenetic data matrix}
\format{
A data frame with 178 observations (characters) on 43 columns
(taxa).
}
\usage{
data("characters")
}
\description{
An example dataset of morphological characters for early tetrapodomorphs
from Simões & Pierce (2021). This type of data would be used as input to
\code{\link[=get_gower_dist]{get_gower_dist()}}.
}
\references{
Simões, T. R. and S. E. Pierce (2021). Sustained High Rates of
Morphological Evolution During the Rise of Tetrapods. \emph{Nature Ecology &
Evolution} 5: 1403–1414.
}
\keyword{datasets}
