% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.dummy.R
\name{drop.dummy}
\alias{drop.dummy}
\alias{drop.dummy.beast}
\alias{drop.dummy.mb}
\title{Remove dummy tip from summary trees, accounting for metadata on the
tips}
\usage{
drop.dummy.beast(
  tree.file,
  output.file = NULL,
  dummy.name = "dummy",
  convert.heights = TRUE
)

drop.dummy.mb(
  tree.file,
  output.file = NULL,
  dummy.name = "dummy",
  convert.ages = TRUE
)
}
\arguments{
\item{tree.file}{path to file containing the tree with dummy tip}

\item{output.file}{path to file to write converted tree. If \code{NULL}
(default), the tree is simply returned.}

\item{dummy.name}{name of the added dummy tip, default \code{dummy}.}

\item{convert.heights, convert.ages}{whether height or age metadata should be converted to height
\itemize{
\item offset or age - offset (required to plot e.g. HPD intervals correctly). Default \code{TRUE}.
}}
}
\value{
A list of \code{tree} converted tree (as treedata) and \code{offset}
height/age of the youngest tip in the final tree
}
\description{
This method is designed to remove the dummy tip added on offset trees once
postprocessing is complete (for instance once the summary tree has been
built using TreeAnnotator). \code{drop.dummy.beast()} is designed to remove the dummy tip added to a dataset before
running with BEAST2, and \code{drop.dummy.mb()} is for running with Mr. Bayes.
}
\examples{
# Analyze the trees with dummy tips - for instance, calculate the MCC summary tree
# Then remove the dummy tip from the MCC tree
final_tree1 <- drop.dummy.beast(system.file("extdata", "ex_offset.MCC.tre", package = "EvoPhylo"))

# Remove the dummy tip from the summary tree
final_tree2 <- drop.dummy.mb(system.file("extdata", "tree_mb_dummy.tre", package = "EvoPhylo"))
}
