% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvoPhylo-package.R
\docType{package}
\name{EvoPhylo-package}
\alias{EvoPhylo}
\alias{EvoPhylo-package}
\title{EvoPhylo: Pre- And Postprocessing of Morphological Data from Relaxed Clock Bayesian Phylogenetics}
\description{
Performs automated morphological character partitioning for phylogenetic analyses and analyze macroevolutionary parameter outputs from clock (time-calibrated) Bayesian inference analyses, following concepts introduced by Simões and Pierce (2021) \doi{10.1038/s41559-021-01532-x}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tiago-simoes/EvoPhylo}
  \item \url{https://tiago-simoes.github.io/EvoPhylo/}
  \item Report bugs at \url{https://github.com/tiago-simoes/EvoPhylo/issues}
}

}
\author{
\strong{Maintainer}: Tiago Simoes \email{trsimoes87@gmail.com} (\href{https://orcid.org/0000-0003-4716-649X}{ORCID})

Authors:
\itemize{
  \item Noah Greifer \email{ngreifer@iq.harvard.edu} (\href{https://orcid.org/0000-0003-3067-7154}{ORCID})
  \item Joelle Barido-Sottani \email{joelle.barido-sottani@m4x.org} (\href{https://orcid.org/0000-0002-5220-5468}{ORCID})
  \item Stephanie Pierce \email{spierce@oeb.harvard.edu} (\href{https://orcid.org/0000-0003-0717-1841}{ORCID})
}

}
\keyword{internal}
