\name{f1}
\alias{f1}

\title{
Clockrate T-tests
}
\description{
Produces a matrixes of 1-sample t-tests for the mean of posterior clockrates against each node's absolute clock rate.
}
\usage{
f1(RatesByClade, posterior.clockrate.all.)
}

\arguments{
  \item{RatesByClade}{
a data frame of rates for different clades, nodes, and clocks. Relative rates should be stored in the \code{rates} column.
}
  \item{posterior.clockrate.all.}{
a vector containing the posterior clock rates (i.e., \code{posterior$clock.rate.}).
}
}
\details{

}
\value{

}

\seealso{

}
\examples{

}
