\name{f3}
\alias{f3}

\title{
Plot clock rate distribution by analysis
}
\description{
Creates a density plot of clock rates by analysis type (tip, tip + node).
}
\usage{
f3(AllRunsrMelted_MC)
}

\arguments{
  \item{AllRunsrMelted_MC}{
(?) Dataset with analysis, Time_bin, and each parameter
}
}
\details{

}
\value{
A \code{ggplot} object, which can be modified using \pkg{ggplot2} functions.
}

\author{

}
\note{

}

\seealso{
\code{\link[ggplot2:geom_density]{ggplot2::geom_density}}
}
\examples{

}
