% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-post_trees.R
\docType{data}
\name{post_trees}
\alias{post_trees}
\title{Multiple phylogenetic clock trees}
\format{
A \code{tidytree} object.
}
\usage{
data("post_trees")
}
\description{
Multiple clock Bayesian phylogenetic tree, imported as an S4 class object
using \code{\link[treeio:beast-parser]{treeio::read.beast()}}.
}
\details{
Example tree file for function \code{\link[=write.beast.treedata]{write.beast.treedata()}}.
}
\seealso{
\code{\link[=write.beast.treedata]{write.beast.treedata()}} for using this file in context.
}
\keyword{datasets}
