% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tree1p.R
\docType{data}
\name{tree1p}
\alias{tree1p}
\title{Phylogenetic tree with a single clock partition}
\format{
A \code{tidytree} object.
}
\usage{
data("tree1p")
}
\description{
A clock Bayesian phylogenetic tree, imported as an S4 class object using
\code{\link[treeio:beast-parser]{treeio::read.mrbayes()}}.
}
\details{
This example tree file was produced by analyzing the data set with a single
morphological partition from Simões & Pierce (2021).
}
\references{
Simões, T. R. and S. E. Pierce (2021). Sustained High Rates of
Morphological Evolution During the Rise of Tetrapods. \emph{Nature Ecology & Evolution} 5: 1403–1414.
}
\seealso{
\code{\link{tree3p}} for another tree object with 3 clock
partitions.

\code{\link[=get_clockrate_table_MrBayes]{get_clockrate_table_MrBayes()}} for extracting the posterior
clockrates from a tree object.
}
\keyword{datasets}
