% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tree_clock1.R
\docType{data}
\name{tree_clock1}
\alias{tree_clock1}
\title{BEAST2 phylogenetic tree with clock rates from partition 1}
\format{
A \code{tidytree} object.
}
\usage{
data("tree_clock1")
}
\description{
A clock Bayesian phylogenetic tree with clock rates from a single clock
partition (partition 1 here), imported as an S4 class object using
\code{\link[treeio:beast-parser]{treeio::read.beast()}}.
}
\details{
This example tree file was produced by analyzing the data set with a single
morphological partition from
}
\seealso{
\code{\link{tree_clock2}} for another BEAST2 tree object with clock rates
from partition 2 for this same dataset.

\code{\link{tree3p}} for another tree object with 3 clock partitions from
Mr.Bayes.

\code{\link{tree1p}} for another tree object with a single clock from
Mr.Bayes.

\code{\link[=get_clockrate_table_BEAST2]{get_clockrate_table_BEAST2()}} for extracting the posterior clock
rates from BEAST2 tree objects.
}
\keyword{datasets}
