\name{tree_clock2}
\alias{tree_clock2}
\docType{data}
\title{
BEAST2 phylogenetic tree with clock rates from partition 2
}
\description{
A clock Bayesian phylogenetic tree with clock rates from a single clock partition (partition 2 here), imported as an S4 class object using \code{treeio::read.beast()}.

}
\usage{data("tree_clock2")}
\format{
A \code{tidytree} object.
}
\details{
This example tree file was produced by analyzing the data set with a single morphological partition from
}
\seealso{

\code{\link{tree_clock1}} for another BEAST2 tree object with  clock rates from partition 1 for this same dataset.

\code{\link{tree3p}} for another tree object with 3 clock partitions from Mr.Bayes.

\code{\link{tree1p}} for another tree object with a single clock from Mr.Bayes.

\code{\link{get_clockrate_table_BEAST2}} for extratcing the poserior clock rates from BEAST2 tree objects.
}
\keyword{datasets}


