\name{characters}
\alias{characters}
\docType{data}
\title{
A morphological phylogenetic data matrix
}
\description{
An example dataset of morphological characters for early tetrapodomorphs from Simões & Pierce (2021). This type of data would be used as input to \code{\link{get_gower_dist}}.
}
\usage{data("characters")}
\format{
  A data frame with 178 observations (characters) on 43 columns (taxa).
}
\keyword{datasets}

\references{
Simões, T. R. and S. E. Pierce (2021). Sustained High Rates of Morphological Evolution During the Rise of Tetrapods. \emph{Nature Ecology & Evolution} 5: 1403–1414.
}
