% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clockrate_table.R
\name{get_clockrate_table}
\alias{get_clockrate_table}
\alias{get_clockrate_table_BEAST2}
\alias{get_clockrate_table_MrBayes}
\title{Extract evolutionary rates from Bayesian clock trees produced by BEAST2 or Mr. Bayes}
\usage{
get_clockrate_table_BEAST2(..., summary = "median", drop_dummy = NULL)

get_clockrate_table_MrBayes(tree, summary = "median", drop_dummy = NULL)
}
\arguments{
\item{...}{\code{treedata} objects containing the summary trees from BEAST2 with associated data on the rates for each separate clock.}

\item{summary}{summary metric used for the rates. Currently supported: \code{"mean"} or \code{"median"}, default \code{"median"}.}

\item{drop_dummy}{if not \code{NULL}, will drop the dummy extant tip with the given label from the summary trees prior to extracting the clock rates (when present). Default is \code{NULL}.}

\item{tree}{An S4 class object of type \code{treedata}; a Bayesian clock tree imported using \code{\link[treeio:beast-parser]{treeio::read.mrbayes()}} for Mr. Bayes summary trees.}
}
\value{
A data frame with a column containing the node identifier (\code{node}) and one column containing the clock rates for each tree provided, in the same order as the trees.
}
\description{
Extract evolutionary rate summary statistics for each node from a Bayesian clock summary tree and stores them in a data frame. BEAST2 stores the rates for each clock in a separate file; all such trees need to be loaded using \code{\link[treeio:beast-parser]{treeio::read.beast()}}.
}
\examples{
#Import all clock summary trees produced by BEAST2 from your local directory
\dontrun{
tree_clock1 <- treeio::read.beast("tree_file_clock1.tre")
tree_clock2 <- treeio::read.beast("tree_file_clock2.tre")
}

#Or use the example BEAST2 multiple clock trees that accompany EvoPhylo.
data(tree_clock1)
data(tree_clock2)

# obtain the rate table from BEAST2 trees
rate_table <- get_clockrate_table_BEAST2(tree_clock1, tree_clock2, summary = "mean")

head(rate_table)

#Import summary tree with three clock partitions produced by
#Mr. Bayes (.t or .tre files) from your local directory
\dontrun{
  tree3p <- treeio::read.mrbayes("Tree3p.t")
}

#Or use the example Mr.Bayes multi-clock tree file (`tree3p`)
data("tree3p")

# obtain the rate table from MrBayes tree
rate_table <- get_clockrate_table_MrBayes(tree3p)

head(rate_table)
}
\seealso{
\code{\link[=clockrate_summary]{clockrate_summary()}} for summarizing and examining properties of the resulting rate table. Note that clade membership for each node must be customized (manually added) before these functions can be used, since this is tree and dataset dependent.

See \code{vignette("rates-selection")} for how to use this function as part of an analysis pipeline.
}
