\name{post_trees}
\alias{post_trees}
\docType{data}
\title{
Multiple phylogenetic clock trees
}
\description{
Multiple clock Bayesian phylogenetic tree, imported as an S4 class object using \code{treeio::read.beast()}.
}
\usage{data("post_trees")}
\format{
A \code{tidytree} object.
}
\details{
Example tree file for function \code{\link{write.beast.treedata}}.
}
\seealso{
\code{\link{write.beast.treedata}} for using this file in context.
}
\keyword{datasets}


