% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tree3p.R
\docType{data}
\name{tree3p}
\alias{tree3p}
\title{Phylogenetic tree with 3 clock partitions}
\format{
A \code{tidytree} object.
}
\description{
A clock Bayesian phylogenetic tree, imported as an S4 class object using
\code{\link[treeio:beast-parser]{treeio::read.mrbayes()}}.
}
\details{
This example tree file was produced by analyzing the data set with 3
morphological clock partitions from Simões & Pierce (2021).
}
\references{
Simões, T. R. and S. E. Pierce (2021). Sustained High Rates of Morphological Evolution During the Rise of Tetrapods. \emph{Nature Ecology & Evolution} 5: 1403–1414.
}
\seealso{
\code{\link{tree1p}} for another tree object with a single clock
partition.

\code{\link[=get_clockrate_table_MrBayes]{get_clockrate_table_MrBayes()}} for extracting the posterior
clockrates from a tree object.
}
\keyword{datasets}
